% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdcate_main_operators.R
\name{HDCATE.set_weight}
\alias{HDCATE.set_weight}
\title{Set weight}
\usage{
HDCATE.set_weight(model, weight = "default")
}
\arguments{
\item{model}{an object created via \link{HDCATE}}

\item{weight}{the vector of sample weights}
}
\value{
None.
}
\description{
When fitting the local linear regression on the second stage, the sample weight can also be manually set  before calling \link{HDCATE.fit}.
}
\examples{
# get simulation data
n_obs <- 500  # Num of observations
n_var <- 100  # Num of observed variables
n_rel_var <- 4  # Num of relevant variables
data <- HDCATE.get_sim_data(n_obs, n_var, n_rel_var)
# conditional expectation model is misspecified
x_formula <- paste(paste0('X', c(2:n_var)), collapse ='+')
# propensity score model is misspecified
# x_formula <- paste(paste0('X', c(1:(n_var-1))), collapse ='+')

# create a new HDCATE model
model <- HDCATE(data=data, y_name='Y', d_name='D', x_formula=x_formula)

# Set user-defined sample weights
HDCATE.set_weight(model, c(rep(0.8, 250), rep(1.2, 250)))
# To reset to default weight, call:
HDCATE.set_weight(model)
}
\keyword{internal}
